Fit base learner

This function fits a selected base learner, which is supposed to be one of random forest, XGBoost, and convolutional neural network, using one of spatiotemporal cross-validation approaches.

base_learner_fit(
  data,
  learner = c("randomforest", "xgboost", "cnn"),
  dependent_name = "pm2.5",
  independent_name = NULL,
  cv_mode = c("lolo", "loto", "lolto", "random", "lblo", "lbto", "lblto"),
  cv_fold = 5L,
  sp_fold = NULL,
  t_fold = NULL,
  blocks = NULL,
  block_id = NULL,
  return_full_object = FALSE,
  ...
)

Arguments

data

stdt. See convert_stobj_to_stdt

learner

character(1). Currently one of "randomforest", "xgboost", and "cnn"

dependent_name

character(1). Name of the dependent variable.

independent_name

character(1). Names of independent variables.

cv_mode

character(1). One of "lolo" (leave-one-location-out), "loto" (leave-one-time-out), "lolto" (leave-one-location-time-out), "lblo" (leave-block-location-out), "lbto" (leave-block-time-out), "lblto" (leave-block-location-time-out), and "random" (full random selection)

cv_fold

integer(1). Number of folds for cross-validation.

sp_fold

integer(1). Number of subfolds for spatial blocks.

t_fold

integer(1). Number of subfolds for temporal blocks.

blocks

integer(2)/sf/SpatVector object.

block_id

character(1). The unique identifier of each block.

return_full_object

logical(1). TRUE will return the object whose class is dependent on the package in use, whereas FALSE will only return the prediction with spatial/temporal information. Full object is potentially helpful for debugging.

...

Arguments passed to base_learner_fit_*.

Value

List of length cv_fold, sp_fold * t_fold, or prod(blocks) (when blocks is a numeric object of length 2) / nrow(blocks) (blocks is sf or SpatVector).

See also

Author

Insang Song