Warp MODIS Swath data into rectilinear grid raster

Swath data is a type of MODIS data organization, where curvilinear points are stored with varying resolution depending on the relative position of the sensor axis. As this type of data typically does not work well with planar spatial data, users should warp or rectify this data into a rectilinear raster. Main procedure is done with stars::st_warp, in which users are able to customize the threshold to fill potential gaps that appear where the target resolution is finer than the local resolution of curvilinear grid points.

modis_warp_stars(path, cellsize = 0.25, threshold = 0.5, crs_out = 4326)

Arguments

path

File path of MODIS swath with exact sub-dataset specification.

cellsize

numeric(1). Cell size (spatial resolution) of output rectilinear grid raster.

threshold

numeric(1). Maximum distance to fill gaps if occur.

crs_out

integer(1)/character(1). Coordinate system definition. Should be compatible with EPSG codes or WKT2. See terra::crs and sf::st_crs / EPSG

Value

stars object.

Note

Users should specify sub-dataset with all flags that are compatible with gdalinfo

See also

Author

Insang Song