Set resource management for SLURM

This function sets up resources for SLURM job submission. Note that this function is designed to work with tar_make_future()

Usage

set_slurm_resource(
  template_file = "inst/targets/template_slurm.tmpl",
  partition = "geo",
  ncpus = 2L,
  ntasks = 2L,
  memory = 8,
  user_email = paste0(Sys.getenv("USER"), "@nih.gov"),
  error_log = "slurm_error.log"
)

Arguments

template_file

SLURM job submission shell template path.

partition

character(1). Name of partition. Default is "geo"

ncpus

integer(1). Number of CPU cores assigned to each task.

ntasks

integer(1). Number of tasks to submit.

memory

integer(1). Specifically odds to 2*x GB.

user_email

character(1). User email address.

error_log

character(1). Error log file name.

Value

A list of resources for tar_resources

Note

This function is designed to be used with tar_resources. Suggested number of ncpus is more than 1 for typical multicore R tasks.

Author

Insang Song