Fit meta learner

This function subsets the full data by column subsamples (rate=50%) The optimal hyperparameter search is performed based on spatial, temporal, and spatiotemporal cross-validation schemes. As of version 0.4.0, the function relies on RMSE to select the best hyperparameter set.

Usage

fit_meta_learner(
  data,
  p_col_sel = 0.5,
  rset = NULL,
  yvar = "Arithmetic.Mean",
  xvar = character(0),
  tune_iter = 50L
)

Arguments

data

data.frame. Full data.

p_col_sel

numeric(1). Rate of column resampling. Default is 0.5.

rset

rset object. Specification of training/test sets.

yvar

character(1). Outcome variable name

xvar

character. Feature names.

tune_iter

integer(1). Bayesian optimization iterations. Default is 50.

Value

List of 3, including the best-fit model, the best hyperparameters, and the all performance records from tune::tune_bayes(). Note that the meta learner function returns the best-fit model, not predicted values.